Refactored util methods into concern
