+++ .rubocop.yml
@@ -27,6 +27,7 @@ require:
 Naming/AsciiIdentifiers:
   Exclude:
     - 'app/controllers/schools/advice/baseload_controller.rb'
+    - 'app/controllers/concerns/advice_pages.rb'
 Security/Eval:
   Exclude:
     - 'app/services/mailchimp_api.rb'
